moved theme provider to another file
