+++ src/index.tsx
@@ -2,11 +2,13 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './index.css'
 import App from './App'
+import { CustomTheme } from './theme/CustomTheme'
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
+    <CustomTheme>
       <App />
-  </React.StrictMode>
-);
-
+    </CustomTheme>
+  </React.StrictMode>,
+)
